Connection and Test adjusted.
